packages/seo-tools/requirements.txt
requests==2.32.3
google-auth==2.39.0
beautifulsoup4==4.13.3
